refactor(data): drop legacy data seeding utilities, replaced by migrations
